fix(core): remove schematic warns on missing project
