+++ packages/workspace/src/schematics/remove/schema.d.ts
@@ -1,4 +1,5 @@
 export interface Schema extends json.JsonObject {
   projectName: string;
-  forceRemove?: boolean;
+  skipFormat: boolean;
+  forceRemove: boolean;
 }
