+++ packages/workspace/src/schematics/remove/schema.json
@@ -29,6 +29,12 @@
       "aliases": ["force-remove"],
       "description": "When true, forces removal even if the project is still in use.",
       "default": false
+    },
+    "skipFormat": {
+      "type": "boolean",
+      "aliases": ["skip-format"],
+      "description": "Skip formatting files.",
+      "default": false
     }
   },
   "required": ["projectName"]
