+++ packages/workspace/src/utils/rules/check-project-exists.ts
@@ -1,15 +1,15 @@
 import { Rule, SchematicContext, Tree } from '@angular-devkit/schematics';
-import { getWorkspace } from '@nrwl/workspace';
 import { from, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { Schema } from '../schema';
+import { getWorkspace } from '../workspace';
 
 /**
  * Checks whether the project exists in the workspace.
+ * Throws an Error if the project is not found.
  *
  * @param schema The options provided to the schematic
  */
-export function checkProjectExists(schema: Schema): Rule {
+export function checkProjectExists(schema: { projectName: string }): Rule {
   return (tree: Tree, _context: SchematicContext): Observable<Tree> => {
     return from(getWorkspace(tree)).pipe(
       map(workspace => {
